Separate builder code into partial class
